fix AcquireNextImageKHR index
Former-commit-id: 5e3c2646dc9914c78644c3c2c62290780f5f2602 [formerly 1a576134406ef26c03dfac6389e2c3aa0dab855e] [formerly ec2c24c42345fbcdb04969149370b4e69d32906f [formerly ec2c24c42345fbcdb04969149370b4e69d32906f [formerly 747b5b8b0801fb502dcec6cf9f5a86b9942f0c3d]]]
Former-commit-id: ec2c24c42345fbcdb04969149370b4e69d32906f
Former-commit-id: 7e4795f22a054919d22dc120e63fe972717a33a9
Former-commit-id: 8349b655c76ae09db76bae8059dc2a370d95877a

+++ src/Rebulk/Renderer/Vulkan/VulkanRenderer.h
@@ -110,7 +110,7 @@ namespace Rbk {
         uint32_t AcquireNextImageKHR(VkSwapchainKHR swapChain, std::pair<std::vector<VkSemaphore>, std::vector<VkSemaphore>>& semaphores);
         void QueueSubmit(VkCommandBuffer commandBuffer);
         void QueueSubmit(uint32_t imageIndex, VkCommandBuffer commandBuffer, std::pair<std::vector<VkSemaphore>, std::vector<VkSemaphore>>& semaphores);
-        size_t QueuePresent(uint32_t imageIndex, VkSwapchainKHR swapChain, std::pair<std::vector<VkSemaphore>, std::vector<VkSemaphore>>& semaphores);
+        uint32_t QueuePresent(uint32_t imageIndex, VkSwapchainKHR swapChain, std::pair<std::vector<VkSemaphore>, std::vector<VkSemaphore>>& semaphores);
         void AddPipelineBarrier(VkCommandBuffer commandBuffer, VkImageMemoryBarrier renderBarrier, VkPipelineStageFlags srcStageMask, VkPipelineStageFlags dstStageMask, VkDependencyFlags dependencyFlags);
         void WaitIdle();
         void GenerateMipmaps(VkCommandBuffer commandBuffer, VkFormat imageFormat, VkImage image, uint32_t texWidth, uint32_t texHeight, uint32_t mipLevels, uint32_t layerCount = 1);
